Se optimizo por completo el codigo, por funcionalidades, componentes, modelos etc. Siguiendo patrones SOLID.
