--- app/src/main/java/com/example/auth/ApiService.kt
@@ -1,4 +1,0 @@
-package com.example.auth
-
-// ApiService movido a com.example.data.auth
-typealias ApiService = com.example.data.auth.ApiService
