--- app/src/main/java/com/example/auth/AuthDataSource.kt
@@ -1,4 +1,0 @@
-package com.example.auth
-
-// AuthDataSource movido a com.example.data.auth – mantener alias para compatibilidad
-typealias AuthDataSource = com.example.data.auth.AuthDataSource
