--- app/src/main/java/com/example/auth/AuthRepository.kt
@@ -1,4 +1,0 @@
-package com.example.auth
-
-// AuthRepository movido a com.example.data.auth
-typealias AuthRepository = com.example.data.auth.AuthRepository
