--- app/src/main/java/com/example/auth/TokenStorage.kt
@@ -1,4 +1,0 @@
-package com.example.auth
-
-// TokenStorage movido a com.example.data.auth
-typealias TokenStorage = com.example.data.auth.TokenStorage
