--- app/src/main/java/com/example/auth/models.kt
@@ -1,7 +1,0 @@
-package com.example.auth
-
-// Modelos movidos a com.example.data.auth – mantenemos aliases para compatibilidad
-typealias RegisterRequest = com.example.data.auth.RegisterRequest
-typealias LoginRequest = com.example.data.auth.LoginRequest
-typealias LoginResponse = com.example.data.auth.LoginResponse
-typealias RefreshResponse = com.example.data.auth.RefreshResponse
