+++ app/src/main/java/com/example/facturacion_inventario/domain/model/Product.kt
@@ -1,0 +1,13 @@
+package com.example.facturacion_inventario.domain.model
+
+import androidx.annotation.DrawableRes
+import com.example.facturacion_inventario.R
+
+// Modelo Product en la capa domain
+data class Product(
+    val id: String,
+    val name: String,
+    val description: String,
+    @DrawableRes val imageRes: Int = R.drawable.ic_motorcycle_animated
+)
+
