--- app/src/main/java/com/example/facturacion_inventario/ui/AppTheme.kt
@@ -1,36 +1,0 @@
-package com.example.facturacion_inventario.ui
-
-import androidx.compose.foundation.isSystemInDarkTheme
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.darkColors
-import androidx.compose.material.lightColors
-import androidx.compose.runtime.Composable
-
-private val LightColorPalette = lightColors(
-    primary = Accent,
-    primaryVariant = PrimaryVariant,
-    secondary = SecondaryColor,
-    background = BackgroundTop,
-    surface = CardColor,
-    onPrimary = OnPrimary,
-    onSecondary = OnSecondary,
-    onBackground = OnBackground,
-    onSurface = OnSurface
-)
-
-private val DarkColorPalette = darkColors(
-    primary = Accent,
-    primaryVariant = PrimaryVariant,
-    secondary = SecondaryColor
-)
-
-@Composable
-fun AppTheme(darkTheme: Boolean = isSystemInDarkTheme(), content: @Composable () -> Unit) {
-    val colors = if (darkTheme) DarkColorPalette else LightColorPalette
-    MaterialTheme(
-        colors = colors,
-        typography = androidx.compose.material.Typography(),
-        shapes = androidx.compose.material.Shapes(),
-        content = content
-    )
-}
